# perf: replace 'input_to_float8' with 'sgl_per_tensor_quant_fp8' (outdated)